+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Pending
 
+* Update stats task
+  * Display a running total of emails processed on stats page
+  * Remove standard deviation stat, it didn't actually mean anything for our dataset
+  * Store join date of oldest user in stats
+  * Calculate disowned inboxes
+  * Calculate `new` & `with_inboxes` via aggregate
 * Update to the latest Salmon (#233)
 * Properly remove Wagtail as a dependency (#254)